Fixed compiler warnings concerning primitive types.   - We are now much stricter concerning comparisons and assignments of     variables of different types (e.g. std::size_t and int).

+++ src/duration_wrapper.hpp
@@ -23,7 +23,8 @@ class duration_wrapper {
 
       // Truncate at 3 decimal points.
       double secs_truncated =
-          ((unsigned int)(d.duration_in_sec_.count() * 1000.0)) / 1000.0;
+          static_cast<std::size_t>(d.duration_in_sec_.count() * 1000.0) /
+          1000.0;
       os << secs_truncated << "s";
 
       return os;
